Moved everything relating to account link management into its own component
